Option to auto-fix icon shading in light themes
Added functions that load, modify, save, and set color scheme files to counteract an issue in which color schemes with light backgrounds automatically darken PNGs, rendering gutter colors inaccurate.

Options are provided to scan & edit global user preferences, syntax-specific settings, and custom files (for users with other plugins affecting the current color scheme). This process only modifies the highest priority file that has set the current scheme, in the order custom, syntax, global, respectively (the `load_settings` command will automatically check defaults for each if no User version is set).

If enabled, the process will be applied to all views (with supported
syntaxes) on plugin load, and to any such view when gaining focus. The
process can also be triggered by a user command (which will force the
scheme to be regenerated even if it already exists).

Modified schemes are saved directly to the Gutter Color package directory
with a special suffix. Although this might clutter the package, it simplifies the color scheme menu.