Force command palette to ignore flags
The palette command 'GutterColor: Fix Color Scheme' was unnecessarily obeying the user's `fix_color_schemes` setting. Therefore, when `false`, as is now default, running the command would do nothing. It makes more sense for this command to be stronger and *always* check global and syntax, because people who don't want the automatic aspect of the fixer should be able to use it to adapt their schemes when necessary.